[#827] Chatplugin uses its own payload
