+++ backend/api/communication/src/main/java/co/airy/core/api/communication/Mapper.java
@@ -1,13 +1,11 @@
 package co.airy.core.api.communication;
 
-import co.airy.avro.communication.Message;
 import co.airy.core.api.communication.dto.Conversation;
 import co.airy.core.api.communication.dto.DisplayName;
-import co.airy.core.api.communication.dto.MessageContainer;
 import co.airy.core.api.communication.payload.ContactResponsePayload;
 import co.airy.core.api.communication.payload.ConversationResponsePayload;
-import co.airy.core.api.communication.payload.MessageResponsePayload;
 import co.airy.model.channel.ChannelPayload;
+import co.airy.model.message.dto.MessageResponsePayload;
 import co.airy.model.metadata.MetadataKeys;
 import org.springframework.stereotype.Component;
 
@@ -15,7 +13,6 @@ import java.util.Map;
 
 import static co.airy.date.format.DateFormat.isoFromMillis;
 import static co.airy.model.metadata.MetadataRepository.getConversationInfo;
-import static co.airy.model.message.MessageRepository.resolveContent;
 
 @Component
 public class Mapper {
@@ -32,7 +29,7 @@ public class Mapper {
                 .tags(conversation.getTagIds())
                 .createdAt(isoFromMillis(conversation.getCreatedAt()))
                 .contact(getContact(conversation))
-                .lastMessage(fromMessageContainer(conversation.getLastMessageContainer()))
+                .lastMessage(MessageResponsePayload.fromMessageContainer(conversation.getLastMessageContainer()))
                 .build();
     }
 
@@ -44,17 +41,6 @@ public class Mapper {
                 .avatarUrl(metadata.get(MetadataKeys.Source.Contact.AVATAR_URL))
                 .displayName(displayName.toString())
                 .info(getConversationInfo(metadata))
-                .build();
-    }
-
-    public MessageResponsePayload fromMessageContainer(MessageContainer messageContainer) {
-        final Message message = messageContainer.getMessage();
-        return MessageResponsePayload.builder()
-                .content(resolveContent(message, messageContainer.getMetadataMap()))
-                .senderType(message.getSenderType().toString().toLowerCase())
-                .deliveryState(message.getDeliveryState().toString().toLowerCase())
-                .id(message.getId())
-                .sentAt(isoFromMillis(message.getSentAt()))
                 .build();
     }
 }
