+++ backend/api/communication/src/main/java/co/airy/core/api/communication/WebSocketController.java
@@ -2,16 +2,17 @@ package co.airy.core.api.communication;
 
 import co.airy.avro.communication.Channel;
 import co.airy.avro.communication.Message;
-import co.airy.model.channel.ChannelPayload;
-import co.airy.core.api.communication.dto.MessageContainer;
 import co.airy.core.api.communication.dto.UnreadCountState;
 import co.airy.core.api.communication.payload.MessageUpsertPayload;
 import co.airy.core.api.communication.payload.UnreadCountPayload;
+import co.airy.model.channel.ChannelPayload;
+import co.airy.model.message.dto.MessageContainer;
+import co.airy.model.message.dto.MessageResponsePayload;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
-import java.util.Map;
 import java.time.Instant;
+import java.util.Map;
 
 import static co.airy.avro.communication.ChannelConnectionState.CONNECTED;
 import static co.airy.date.format.DateFormat.isoFromMillis;
@@ -36,7 +37,7 @@ public class WebSocketController {
         final MessageUpsertPayload messageUpsertPayload = MessageUpsertPayload.builder()
                 .channelId(message.getChannelId())
                 .conversationId(message.getConversationId())
-                .message(mapper.fromMessageContainer(new MessageContainer(message, Map.of())))
+                .message(MessageResponsePayload.fromMessageContainer(new MessageContainer(message, Map.of())))
                 .build();
         messagingTemplate.convertAndSend(QUEUE_MESSAGE, messageUpsertPayload);
     }
