--- backend/api/communication/src/main/java/co/airy/core/api/communication/payload/MessageResponsePayload.java
@@ -1,18 +1,0 @@
-package co.airy.core.api.communication.payload;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageResponsePayload {
-    private String id;
-    private String content;
-    private String senderType;
-    private String sentAt;
-    private String deliveryState;
-}
