+++ backend/api/communication/src/main/java/co/airy/core/api/communication/payload/MessageUpsertPayload.java
@@ -1,5 +1,6 @@
 package co.airy.core.api.communication.payload;
 
+import co.airy.model.message.dto.MessageResponsePayload;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
