+++ backend/model/message/BUILD
@@ -9,6 +9,8 @@ custom_java_library(
         "//lib/java/kafka/schema:application-communication-messages",
     ],
     deps = [
+        "//:lombok",
+        "//lib/java/date",
         "//backend/avro/communication:message-avro",
     ],
 )
