+++ backend/model/message/src/main/java/co/airy/model/message/MessageMapper.java
@@ -1,0 +1,4 @@
+package co.airy.model.message;
+
+public class MessageMapper {
+}
