+++ backend/sources/chat-plugin/src/main/java/co/airy/core/chat_plugin/WebSocketController.java
@@ -2,6 +2,7 @@ package co.airy.core.chat_plugin;
 
 import co.airy.avro.communication.Message;
 import co.airy.core.chat_plugin.payload.MessageUpsertPayload;
+import co.airy.model.message.dto.MessageResponsePayload;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
@@ -9,11 +10,9 @@ import org.springframework.stereotype.Service;
 public class WebSocketController {
     public static final String QUEUE_MESSAGE = "/queue/message";
     private final SimpMessagingTemplate messagingTemplate;
-    private final Mapper mapper;
 
-    WebSocketController(SimpMessagingTemplate messagingTemplate, Mapper mapper) {
+    WebSocketController(SimpMessagingTemplate messagingTemplate) {
         this.messagingTemplate = messagingTemplate;
-        this.mapper = mapper;
     }
 
     public void onNewMessage(Message message) {
@@ -21,7 +20,7 @@ public class WebSocketController {
                 MessageUpsertPayload.builder()
                         .channelId(message.getChannelId())
                         .conversationId(message.getConversationId())
-                        .message(mapper.fromMessage(message))
+                        .message(MessageResponsePayload.fromMessage(message))
                         .build());
     }
 }
