+++ backend/sources/chat-plugin/src/main/java/co/airy/core/chat_plugin/payload/AuthenticationResponsePayload.java
@@ -1,5 +1,6 @@
 package co.airy.core.chat_plugin.payload;
 
+import co.airy.model.message.dto.MessageResponsePayload;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
